📦 refactor: reorganize repositories/ into index/repository files
- Rename `<domain>.repository.ts` to `repository.ts` in each `<domain>-repository/` folder, mirroring the `models/` convention
- Split `status-repository`'s inline logic out of `index.ts` into its own `repository.ts`, with `index.ts` reduced to a barrel export
