+++ src/repositories/user-activation-token-repository/index.ts
@@ -1,2 +1,2 @@
+export * from './repository'
 export * from './types'
-export * from './user-activation-token.repository'
